+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.4.3
 
 ##### Add
-* [issue#58] Add `publishSocketTimeoutInSeconds` param to change the socket timeout for publish requests in seconds.
+* [issue#58] Add new `publishSocketTimeoutInSeconds` parameter to change the socket timeout for publish requests in seconds.
+* [issue#59] Add new `credentials` parameter to provide credentials as a base64 encoded string.
 
 ##### Breaking Changes
 * Remove support of Sonatype. It means that you can't use the plugin from Maven Central. You must to use the Gradle Portal.
@@ -13,6 +14,7 @@
       }
   }
   ```
+* Remove support of `clientId` and `clientSecret` CLI params. Use `credentials` param instead.
 
 # 1.4.2